# frontend/trufotbot-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "trufotbot",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/patients/{patient_id}/medications/{medication_id}/remind": {
      "put": {
        "tags": [
          "reminders"
        ],
        "summary": "Send a reminder",
        "operationId": "reminders_send",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "description": "Patient ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "medication_id",
            "in": "path",
            "description": "Medication ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reminder sent successfully"
          },
          "404": {
            "description": "Medication not found"
          }
        }
      }
    },
    "/api/patients/{patient_id}/medications/{medication_id}/reminders": {
      "get": {
        "tags": [
          "reminders"
        ],
        "summary": "Get reminders for patient's medication",
        "operationId": "reminders_get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "description": "Patient ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "medication_id",
            "in": "path",
            "description": "Medication ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reminders fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reminders"
                }
              }
            }
          },
          "404": {
            "description": "Medication not found"
          }
        }
      },
      "put": {
        "tags": [
          "reminders"
        ],
        "summary": "Set reminders for patient's medication",
        "operationId": "reminders_set",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "description": "Patient ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "medication_id",
            "in": "path",
            "description": "Medication ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Reminders"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Reminders set successfully"
          },
          "404": {
            "description": "Medication not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Reminders": {
        "type": "object",
        "required": [
          "cron_schedules"
        ],
        "properties": {
          "cron_schedules": {
            "type": "array",
            "items": {
              "type": "string",
              "examples": [
                "4:1,24:3"
              ]
            }
          }
        }
      },
...
    }
  },
...
}