# src/handlers/reminders.rs
// Copyright (C) 2026 Ohad Lutzky <lutzky@gmail.com>
//
// SPDX-License-Identifier: GPL-3.0-only

use std::sync::Arc;

use axum::{
    Json,
    extract::{Path, State},
};
use chrono::{DateTime, Utc};
use color_eyre::eyre::{self, eyre};
use teloxide::utils::markdown;

use crate::{
    api::patient::Reminders,
    app_state::Config,
    errors::ServiceError,
    ids::{MedicationId, MessageId, PatientId},
    messenger::{Messenger, callbacks},
    models::{self, Medication},
    reminder_scheduler::ReminderScheduler,
    storage::Storage,
    time::now,
};

pub const UTOIPA_TAG: &str = "reminders";

fn validate_cron_schedule(schedule: &str) -> Result<(), ServiceError> {
    // Although it would be nice to use [tokio_cron_scheduler::Job::schedule_to_cron] to validate
    // this, validation currently (2025-11-02) only works with feature `english` enabled, and
    // otherwise does nothing.
    #[allow(clippy::unreachable)] // This is never scheduled
    tokio_cron_scheduler::Job::new(schedule, |_, _| unreachable!())
        .map_err(|_| ServiceError::bad_request(format!("Invalid cron schedule '{schedule}'")))?;
    Ok(())
}

#[utoipa::path(
    get,
    path = "/api/patients/{patient_id}/medications/{medication_id}/reminders",
    summary = "Get reminders for patient's medication",
    operation_id = "reminders_get",
    tag = UTOIPA_TAG,
    responses(
        (status = 200, description = "Reminders fetched successfully", body=Reminders),
        (status = 404, description = "Medication not found"),
    ),
    params(
        ("patient_id" = i32, Path, description = "Patient ID"),
        ("medication_id" = i32, Path, description = "Medication ID"),
    )
)]
pub async fn get(
    State(storage): State<Storage>,
    Path((patient_id, medication_id)): Path<(PatientId, MedicationId)>,
) -> Result<Json<Reminders>, ServiceError> {
    struct ReminderRow {
        cron_schedule_lines: String,
    }

    let schedules = sqlx::query_as!(
        ReminderRow,
        r#"
        SELECT
            r.cron_schedule AS "cron_schedule_lines!"
        FROM reminders r
        WHERE r.patient_id = ?
          AND r.medication_id = ?
        "#,
        patient_id,
        medication_id,
    )
    .fetch_optional(&storage.pool)
    .await?;

    let cron_schedules = match schedules {
        Some(s) => s.cron_schedule_lines.lines().map(String::from).collect(),
        None => vec![],
    };

    Ok(Json(Reminders { cron_schedules }))
}

#[utoipa::path(
    put,
    path = "/api/patients/{patient_id}/medications/{medication_id}/reminders",
    summary = "Set reminders for patient's medication",
    operation_id = "reminders_set",
    tag = UTOIPA_TAG,
    responses(
        (status = 200, description = "Reminders set successfully"),
        (status = 404, description = "Medication not found"),
    ),
    request_body = Reminders,
    params(
        ("patient_id" = i32, Path, description = "Patient ID"),
        ("medication_id" = i32, Path, description = "Medication ID"),
    )
)]
pub async fn set(
    State(storage): State<Storage>,
    State(mut reminder_scheduler): State<ReminderScheduler>,
    Path((patient_id, medication_id)): Path<(PatientId, MedicationId)>,
    Json(Reminders { cron_schedules }): Json<Reminders>,
) -> Result<(), ServiceError> {
    for schedule in &cron_schedules {
        validate_cron_schedule(schedule)?;
    }

    let joined_cron_schedule = cron_schedules.join("\n");

    sqlx::query_as!(
        ReminderRow,
        r#"
        INSERT INTO reminders (patient_id, medication_id, cron_schedule)
        VALUES (?, ?, ?)
        ON CONFLICT (patient_id, medication_id) DO UPDATE
        SET cron_schedule = EXCLUDED.cron_schedule
        "#,
        patient_id,
        medication_id,
        joined_cron_schedule,
    )
    .execute(&storage.pool)
    .await?;

    let cron_schedules_ref: Vec<&str> = cron_schedules.iter().map(|s| s.as_str()).collect();

    reminder_scheduler
        .set_reminders(patient_id, medication_id, &cron_schedules_ref)
        .await?;

    Ok(())
}

#[utoipa::path(
    put,
    path = "/api/patients/{patient_id}/medications/{medication_id}/remind",
    summary = "Send a reminder",
    operation_id = "reminders_send",
    tag = UTOIPA_TAG,
    responses(
        (status = 200, description = "Reminder sent successfully"),
        (status = 404, description = "Medication not found"),
    ),
    params(
        ("patient_id" = i32, Path, description = "Patient ID"),
        ("medication_id" = i32, Path, description = "Medication ID"),
    )
)]
pub async fn send_reminder(
    State(storage): State<Storage>,
    State(messenger): State<Messenger>,
    State(config): State<Arc<Config>>,
    Path((patient_id, medication_id)): Path<(PatientId, MedicationId)>,
) -> Result<(), ServiceError> {
    let patient = models::Patient::get(&storage.pool, patient_id).await?;
    let medication = models::Medication::get(&storage.pool, medication_id).await?;
    let latest_dosage = Medication::latest_dosage(&storage.pool, medication_id, patient_id).await?;

    if patient.telegram_group_id.is_none() {
        return Err(ServiceError::bad_request(
            "Patient has no telegram group ID",
        ));
    }

    let default_dosage = latest_dosage.unwrap_or(1.0);

    let base_message = markdown::escape(&format!(
        "🔔 Time for {} to take {}.",
        patient.name, medication.name
    ));

    let message_id = messenger
        .send(&patient, base_message.clone(), vec![])
        .await?
        .ok_or_else(|| {
            ServiceError::InternalError(eyre!(
                "Sending message to patient {patient_id:?} returned None, \
                 though we checked that they have a telegram group ID"
            ))
        })?;

    let url = deep_link(patient_id, medication_id, message_id.id(), now(), &config);

    let keyboard = vec![
        Some((
            format!("Take {default_dosage} 💊"),
            callbacks::Action::TakeFromReminder {
                patient_id,
                medication_id,
                quantity: default_dosage,
            },
        )),
        Some((
            "Skip ⏭️".to_string(),
            callbacks::Action::TakeFromReminder {
                patient_id,
                medication_id,
                quantity: 0.0,
            },
        )),
        match url {
            Ok(url) => Some(("Take... 📝".to_string(), callbacks::Action::Link { url })),
            Err(err) => {
                log::error!("Couldn't build URL to display \"Take...\" button: {err}");
                None
            }
        },
    ]
    .into_iter()
    .flatten()
    .collect::<Vec<_>>();

    messenger
        .edit(&patient, None, message_id.id(), base_message, keyboard)
        .await?;

    Ok(())
}

fn deep_link(
    patient_id: PatientId,
    medication_id: MedicationId,
    message_id: MessageId,
    reminder_sent_time: DateTime<Utc>,
    config: &Config,
) -> eyre::Result<url::Url> {
    let mut url = config.frontend_url.clone();

    url.path_segments_mut()
        .map_err(|_| {
            // path_segments_mut() returns a Result<_, ()>, so there's nothing to wrap
            eyre!(
                "frontend_url {:?} can't be used as a base",
                config.frontend_url
            )
        })?
        .push("patients")
        .push(&patient_id.to_string())
        .push("medications")
        .push(&medication_id.to_string());

    url.query_pairs_mut()
        .append_pair("message_id", &message_id.to_string())
        .append_pair("message_time", &reminder_sent_time.timestamp().to_string())
        .finish();

    Ok(url)
}

#[cfg(test)]
mod tests {
    use std::sync::Arc;

    use crate::{
        api::{dose, requests::CreateDoseQueryParams},
        app_state::Config,
        ids::DoseId,
        time::FAKE_TIME,
    };
    use axum::{Json, extract::Query};

    use pretty_assertions::assert_eq;
    use sqlx::SqlitePool;

    use super::*;

    use crate::test_utils::{dose_keyboard, dt, md, reminder_url};
    use crate::{
        app_state::AppState,
        messenger::fake_sender::{FakeSender, messages_from_slice},
    };

    async fn setup(db: SqlitePool, config: Arc<Config>) -> (AppState, Arc<FakeSender>) {
        let fake_telegram = Arc::new(FakeSender::new());
        let messenger = fake_telegram.clone().into();
        let app_state = AppState::new(db, messenger, config).await.unwrap();
        (app_state, fake_telegram)
    }

    async fn test_remind_dose(db: SqlitePool, delete_and_resend: bool, quantity: f64) {
        let config = Arc::new(Config {
            trufotbot_reminder_completion_delete_and_resend: delete_and_resend,
            trufotbot_show_dose_absolute_time: true,
            ..Config::load().unwrap()
        });
        let frontend_url = config.frontend_url.clone();
        let (app_state, fake_telegram) = setup(db, config).await;

        FAKE_TIME
            .scope("2025-01-02T00:00:00Z", async {
                send_reminder(
                    State(app_state.storage.clone()),
                    State(app_state.messenger.clone()),
                    State(app_state.config.clone()),
                    Path((PatientId(1), MedicationId(1))),
                )
                .await
                .unwrap();
            })
            .await;

        let reminder_url =
            reminder_url(&frontend_url, PatientId(1), MedicationId(1), 1, 1735776000);
        let keyboard = [
            (
                "Take 1 💊",
                callbacks::Action::TakeFromReminder {
                    patient_id: PatientId(1),
                    medication_id: MedicationId(1),
                    quantity: 1.0,
                },
            ),
            (
                "Skip ⏭️",
                callbacks::Action::TakeFromReminder {
                    patient_id: PatientId(1),
                    medication_id: MedicationId(1),
                    quantity: 0.0,
                },
            ),
            ("Take... 📝", callbacks::Action::Link { url: reminder_url }),
        ];

        assert_eq!(
            fake_telegram.messages.get_messages(-123).await.unwrap(),
            messages_from_slice(&[(&md("🔔 Time for Alice to take Aspirin."), &keyboard)], 1)
        );

        let taken_at = dt("2025-01-01T23:00:00Z");
        let reminded_at = dt("2025-01-01T22:00:00Z");

        FAKE_TIME
            .scope("2025-01-02T00:00:00Z", async {
                crate::handlers::doses::record(
                    Path((PatientId(1), MedicationId(1))),
                    Query(CreateDoseQueryParams {
                        reminder_message_id: Some(MessageId(1)),
                        reminder_sent_time: Some(reminded_at),
                    }),
                    State(app_state.storage.clone()),
                    State(app_state.messenger.clone()),
                    State(app_state.config.clone()),
                    Json(dose::CreateDose {
                        quantity,
                        taken_at,
                        noted_by_user: Some("Albert".to_string()),
                    }),
                )
                .await
                .unwrap();
            })
            .await;

        let (expected_text, expected_id) = if delete_and_resend {
            (
                md(&if quantity == 0.0 {
                    concat!(
                        "⏭️ Albert decided to skip giving Alice Aspirin (0) an hour earlier ",
                        "(2025-01-01 (Wed) 23:00)"
                    )
                    .to_string()
                } else {
                    format!(
                        concat!(
                            "✅ Albert gave Alice Aspirin ({quantity})",
                            " an hour earlier (2025-01-01 (Wed) 23:00)",
                        ),
                        quantity = quantity,
                    )
                }),
                2,
            )
        } else {
            (
                md(&if quantity == 0.0 {
                    concat!(
                        "⏭️ Albert decided to skip giving Alice Aspirin (0) an hour later ",
                        "(2025-01-01 (Wed) 23:00)"
                    )
                    .to_string()
                } else {
                    format!(
                        concat!(
                            "✅ Albert gave Alice Aspirin ({quantity})",
                            " an hour later (2025-01-01 (Wed) 23:00)",
                        ),
                        quantity = quantity,
                    )
                }),
                1,
            )
        };

        assert_eq!(
            fake_telegram.messages.get_messages(-123).await.unwrap(),
            messages_from_slice(
                &[(
                    &expected_text,
                    &dose_keyboard(
                        PatientId(1),
                        MedicationId(1),
                        DoseId(1),
                        quantity,
                        &frontend_url
                    ),
                )],
                expected_id
            )
        );
    }

    #[sqlx::test(fixtures("../fixtures/patients.sql", "../fixtures/medications.sql"))]
    async fn remind_dose_succeeds_with_edit(db: SqlitePool) {
        test_remind_dose(db, false, 2.0).await;
    }

    #[sqlx::test(fixtures("../fixtures/patients.sql", "../fixtures/medications.sql"))]
    async fn remind_dose_succeeds_with_delete_and_resend(db: SqlitePool) {
        test_remind_dose(db, true, 2.0).await;
    }

    #[sqlx::test(fixtures("../fixtures/patients.sql", "../fixtures/medications.sql"))]
    async fn remind_dose_skip_succeeds_with_edit(db: SqlitePool) {
        test_remind_dose(db, false, 0.0).await;
    }

    #[sqlx::test(fixtures("../fixtures/patients.sql", "../fixtures/medications.sql"))]
    async fn remind_dose_skip_succeeds_with_delete_and_resend(db: SqlitePool) {
        test_remind_dose(db, true, 0.0).await;
    }

    #[sqlx::test(fixtures("../fixtures/patients.sql", "../fixtures/medications.sql"))]
    async fn remind_dose_then_edit_succeeds_with_delete_and_resend(db: SqlitePool) {
        let config = Arc::new(Config {
            trufotbot_reminder_completion_delete_and_resend: true,
            trufotbot_show_dose_absolute_time: true,
            ..Config::load().unwrap()
        });
        let frontend_url = config.frontend_url.clone();
        let (app_state, fake_telegram) = setup(db, config.clone()).await;

        let taken_at = dt("2025-01-02T00:00:00Z");
        let reminded_at = dt("2025-01-01T23:00:00Z");

        // Send reminder at FAKE_TIME = 2025-01-02T00:00:00Z
        FAKE_TIME
            .scope("2025-01-02T00:00:00Z", async {
                send_reminder(
                    State(app_state.storage.clone()),
                    State(app_state.messenger.clone()),
                    State(config.clone()),
                    Path((PatientId(1), MedicationId(1))),
                )
                .await
                .unwrap();
            })
            .await;

        // Record dose from reminder (with reminded_at = 1 hour before now, taken_at = now)
        FAKE_TIME
            .scope("2025-01-02T00:00:00Z", async {
                crate::handlers::doses::record(
                    Path((PatientId(1), MedicationId(1))),
                    Query(CreateDoseQueryParams {
                        reminder_message_id: Some(MessageId(1)),
                        reminder_sent_time: Some(reminded_at),
                    }),
                    State(app_state.storage.clone()),
                    State(app_state.messenger.clone()),
                    State(config.clone()),
                    Json(dose::CreateDose {
                        quantity: 2.0,
                        taken_at,
                        noted_by_user: Some("Albert".to_string()),
                    }),
                )
                .await
                .unwrap();
            })
            .await;

        // Verify initial message says "now" (message_time = taken_at = now())
        assert_eq!(
            fake_telegram.messages.get_messages(-123).await.unwrap(),
            messages_from_slice(
                &[(
                    &md("✅ Albert gave Alice Aspirin (2) now (2025-01-02 (Thu) 00:00)"),
                    &dose_keyboard(PatientId(1), MedicationId(1), DoseId(1), 2.0, &frontend_url),
                )],
                2
            )
        );

        // Update the dose (edit) at FAKE_TIME = 2025-01-02T00:05:00Z
        FAKE_TIME
            .scope("2025-01-02T00:05:00Z", async {
                crate::handlers::doses::update(
                    Path((PatientId(1), MedicationId(1), DoseId(1))),
                    State(app_state.messenger.clone()),
                    State(app_state.storage.clone()),
                    State(config.clone()),
                    Json(dose::CreateDose {
                        quantity: 1.0,
                        taken_at,
                        noted_by_user: Some("Bob".to_string()),
                    }),
                )
                .await
                .unwrap();
            })
            .await;

        // Verify edited message still says "now" (DB-stored telegram_message_time = now(), not
        // reminded_at).
        assert_eq!(
            fake_telegram.messages.get_messages(-123).await.unwrap(),
            messages_from_slice(
                &[(
                    &md("✏️✅ Bob gave Alice Aspirin (1) now (2025-01-02 (Thu) 00:00)"),
                    &dose_keyboard(PatientId(1), MedicationId(1), DoseId(1), 1.0, &frontend_url),
                )],
                2
            )
        );
    }

    #[sqlx::test(fixtures("../fixtures/patients.sql", "../fixtures/medications.sql"))]
    async fn remind_dose_then_edit_to_skip_succeeds_with_delete_and_resend(db: SqlitePool) {
        let config = Arc::new(Config {
            trufotbot_reminder_completion_delete_and_resend: true,
            trufotbot_show_dose_absolute_time: true,
            ..Config::load().unwrap()
        });
        let frontend_url = config.frontend_url.clone();
        let (app_state, fake_telegram) = setup(db, config.clone()).await;

        let taken_at = dt("2025-01-02T00:00:00Z");
        let reminded_at = dt("2025-01-01T23:00:00Z");

        // Send reminder at FAKE_TIME = 2025-01-02T00:00:00Z
        FAKE_TIME
            .scope("2025-01-02T00:00:00Z", async {
                send_reminder(
                    State(app_state.storage.clone()),
                    State(app_state.messenger.clone()),
                    State(config.clone()),
                    Path((PatientId(1), MedicationId(1))),
                )
                .await
                .unwrap();
            })
            .await;

        // Record dose from reminder (with reminded_at = 1 hour before now, taken_at = now)
        FAKE_TIME
            .scope("2025-01-02T00:00:00Z", async {
                crate::handlers::doses::record(
                    Path((PatientId(1), MedicationId(1))),
                    Query(CreateDoseQueryParams {
                        reminder_message_id: Some(MessageId(1)),
                        reminder_sent_time: Some(reminded_at),
                    }),
                    State(app_state.storage.clone()),
                    State(app_state.messenger.clone()),
                    State(config.clone()),
                    Json(dose::CreateDose {
                        quantity: 2.0,
                        taken_at,
                        noted_by_user: Some("Albert".to_string()),
                    }),
                )
                .await
                .unwrap();
            })
            .await;

        // Verify initial message says "now" (message_time = taken_at = now())
        assert_eq!(
            fake_telegram.messages.get_messages(-123).await.unwrap(),
            messages_from_slice(
                &[(
                    &md("✅ Albert gave Alice Aspirin (2) now (2025-01-02 (Thu) 00:00)"),
                    &dose_keyboard(PatientId(1), MedicationId(1), DoseId(1), 2.0, &frontend_url),
                )],
                2
            )
        );

        // Update the dose (edit) at FAKE_TIME = 2025-01-02T00:05:00Z, skipping it
        FAKE_TIME
            .scope("2025-01-02T00:05:00Z", async {
                crate::handlers::doses::update(
                    Path((PatientId(1), MedicationId(1), DoseId(1))),
                    State(app_state.messenger.clone()),
                    State(app_state.storage.clone()),
                    State(config.clone()),
                    Json(dose::CreateDose {
                        quantity: 0.0,
                        taken_at,
                        noted_by_user: Some("Bob".to_string()),
                    }),
                )
                .await
                .unwrap();
            })
            .await;

        // Verify edited message still says "now" (DB-stored telegram_message_time = now(), not
        // reminded_at).
        assert_eq!(
            fake_telegram.messages.get_messages(-123).await.unwrap(),
            messages_from_slice(
                &[(
                    &md(concat!(
                        "✏️⏭️ Bob decided to skip giving Alice Aspirin (0) now ",
                        "(2025-01-02 (Thu) 00:00)"
                    )),
                    &dose_keyboard(PatientId(1), MedicationId(1), DoseId(1), 0.0, &frontend_url),
                )],
                2
            )
        );
    }
}